move test version of build_query to test code
