+++ crates/vm/src/program.rs
@@ -3,10 +3,7 @@
 //! It carries an [EnvDb] with the functions, idents, types.
 
 use crate::errors::ErrorVm;
-use crate::eval::{build_query, build_source_expr_query};
 use crate::expr::{Code, CrudExpr, SourceSet};
-use crate::rel_ops::RelOps;
-use crate::relation::MemTable;
 use spacetimedb_sats::ProductValue;
 
 /// A trait to allow split the execution of `programs` to allow executing
@@ -22,42 +19,3 @@ pub trait ProgramVm {
 }
 
 pub type Sources<'a, const N: usize> = &'a mut SourceSet<Vec<ProductValue>, N>;
-
-/// A default program that run in-memory without a database
-pub struct Program;
-
-impl ProgramVm for Program {
-    fn eval_query<const N: usize>(&mut self, query: CrudExpr, sources: Sources<'_, N>) -> Result<Code, ErrorVm> {
-        match query {
-            CrudExpr::Query(query) => {
-                let result = build_source_expr_query(sources, &query.source);
-                let rows = build_query(result, &query.query, sources)?.collect_vec(|row| row.into_product_value());
-
-                let head = query.head().clone();
-
-                Ok(Code::Table(MemTable::new(head, query.source.table_access(), rows)))
-            }
-            CrudExpr::Insert { .. } => {
-                todo!()
-            }
-            CrudExpr::Update { .. } => {
-                todo!()
-            }
-            CrudExpr::Delete { .. } => {
-                todo!()
-            }
-            CrudExpr::CreateTable { .. } => {
-                todo!()
-            }
-            CrudExpr::Drop { .. } => {
-                todo!()
-            }
-            CrudExpr::SetVar { .. } => {
-                todo!()
-            }
-            CrudExpr::ReadVar { .. } => {
-                todo!()
-            }
-        }
-    }
-}
